factor_model_descriptive/requirements_des.txt
SQLAlchemy==1.3.18
pandas
numpy
scikit-learn==1.0.0
python-dateutil
lightgbm
psycopg2-binary
hyperopt
xarray
matplotlib
openpyxl
s-dbw
fuzzy-c-means
slack_sdk
pyclustertend
statsmodels
firebase_admin
